Add zipcode verification in our rust code to parse and query KML files
We parse the location based KMZ files (a compressed KML file) to load zipcode sectors into memory
then we parse these zipcode sectors into relevant geometry types so we cna query if a lat lng lies within them

This should allow an okay-ish free version of reverse geocoding to validate geocoding from services like google which happen to be innacurate sometimes.

This functionality is then wrapped in the newtype pattern with refcell to allow for interior mutability, bypassing magnus' constraints with method! which allow only for &self and not &mut self on method args
We expose the newtype through a ruby interface to allow initialising and querying of our rust struct